+++ classes/Data/Task/class.LogEntry.php
@@ -18,6 +18,7 @@ class LogEntry extends RecordData
     const TYPE_NOTE = 'note';
     const TYPE_WORKING_TIME_CHANGE = 'working_time_change';
     const TYPE_WORKING_TIME_DELETE = 'working_time_delete';
+    const TYPE_WRITER_NOTE = 'writer_note';
     const TYPE_WRITER_REMOVAL = 'writer_removal';
     const TYPE_WRITER_EXCLUSION = 'writer_exclusion';
     const TYPE_WRITER_REPEAL_EXCLUSION = 'writer_repeal_exclusion';
@@ -30,6 +31,7 @@ class LogEntry extends RecordData
         self::TYPE_NOTE => self::CATEGORY_NOTE,
         self::TYPE_WORKING_TIME_CHANGE => self::CATEGORY_WORKING_TIME,
         self::TYPE_WORKING_TIME_DELETE => self::CATEGORY_WORKING_TIME,
+        self::TYPE_WRITER_NOTE => self::CATEGORY_NOTE,
         self::TYPE_WRITER_REMOVAL => self::CATEGORY_EXCLUSION,
         self::TYPE_WRITER_EXCLUSION => self::CATEGORY_EXCLUSION,
         self::TYPE_WRITER_REPEAL_EXCLUSION => self::CATEGORY_EXCLUSION,